+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 3.0.0 (unreleased)
 
+* `logging` – Remove deprecated plugin (@janko)
+
 * `s3` – Remove deprecated `S3#download` method (@janko)
 
 * `s3` – Remove deprecated `S3#stream` method (@janko)